requirements.txt
astreamlit>=1.28.0
torch>=2.0.0
pydub>=0.25.1
openai-whisper>=20231117
tqdm>=4.65.0
requests>=2.31.0
faster-whisper>=1.1.1
selenium==4.29.0
psutil>=5.9.0
fpdf2>=2.7.0
